backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.42
aiosqlite==0.20.0
greenlet==3.2.4
redis==5.2.1
pydantic-settings==2.10.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
openai==1.99.9
httpx==0.28.1
